Moved MarketingConsentCard to a different class. Removed width and height attributes.
